Avoid open redirect and preserve query string
CodeQL reported a potential server-side URL redirect vulnerability due
to the use of `ctx.url.toLowerCase()`. In particular, a schema-less URL
such as `//evil.com/HELLO` could result in a redirect to
`//evil.com/hello` which browsers interpret as a cross-origin redirect,
enabling potential phishing attacks.

Additionally, the middleware did not preserve query strings during
redirection which is probably not the desired behavior in most
scenarios.

Update the middleware to parse the URL using the current host as a base,
lowercase only the pathname, and preserve the query string as-is. This
ensures all redirects remain safe, relative paths and preserves the
original intent of the request.